requirements.txt
pyserial==3.4
imageio==2.6.1
pigpio==1.46